bot/requirements.txt
nextcord==2.6.0
aiohttp==3.9.5
python-dotenv==1.0.1
